+++ lib/shadcn_ui/components/foundation/alert.ex
@@ -1,0 +1,3 @@
+defmodule ShadcnUI.Components.Foundation.Alert do
+  @moduledoc "Public Alert function components."
+end