+++ lib/shadcn_ui/components/foundation/avatar.ex
@@ -1,0 +1,3 @@
+defmodule ShadcnUI.Components.Foundation.Avatar do
+  @moduledoc "Public Avatar function components."
+end