+++ lib/shadcn_ui/components/foundation/skeleton.ex
@@ -1,0 +1,3 @@
+defmodule ShadcnUI.Components.Foundation.Skeleton do
+  @moduledoc "Public Skeleton function components."
+end