+++ test/shadcn_ui_test.exs
@@ -1,7 +1,0 @@
-defmodule ShadcnUITest do
-  use ExUnit.Case
-
-  test "defines the package entry module" do
-    assert Code.ensure_loaded?(ShadcnUI)
-  end
-end